move compat handlers under fml/compat, register extra transformers as strings
